made feed v2 default
