+++ Modules/feed/feed_controller.php
@@ -35,13 +35,7 @@ function feed_controller()
 
         textdomain("messages");
         if ($route->action == "list" && $session['write']) {
-            $ui_version_2 = $user->get_preferences($session['userid'], 'deviceView');
-
-            if (isset($ui_version_2) && $ui_version_2) {
-                return view("Modules/feed/Views/feedlist_view_v2.php",array());
-            } else {
-                return view("Modules/feed/Views/feedlist_view.php",array());
-            }
+            return view("Modules/feed/Views/feedlist_view.php");
         }
         else if ($route->action == "api" && $session['write']) return view("Modules/feed/Views/feedapi_view.php",array());
         else if (!$session['read']) return ''; // empty strings force user back to login
